requirements.txt
fastapi==0.112.2
uvicorn==0.30.5
passlib
python-jose[cryptography]
pydantic[email]
